Synchronize creation of cacheEntries for existing ids
When creating cache entries for already existing (but not yet cached)
sessions, make sure only one entry is created and only one delegate
sessions is fetched.

In order to keep locking to a minium, a separate lock is used for every
session id.

For creation of new entries this should not be needed, even though it
populates the session cache, because the id of the new session is not
known to any client until the entry is in the cache.